Address expert review feedback on the MTP hardening
Localize the missing-uid error. The message reaches the user verbatim -
StartTestRun funnels ex.Message into HandleLogMessage(Error) - and every other
user-facing TestPlatformException in this assembly is resourced, so a hardcoded
English string formatted with CurrentCulture was self-contradictory. Adds
MtpTestCaseMissingNodeUid to Resources.resx, the generated designer property,
and a trans-unit to all 13 xlf files. The text now also states the remedy
(re-run discovery, or run without a selection) rather than only naming the
failure, and the comment records that aborting the whole source is deliberate:
silently running the addressable subset would recreate the same class of bug in
a smaller form.

Mark the three new test classes DoNotParallelize. MSTest parallelizes across
classes at MethodLevel by default here, and these classes mutate process-global
state - the MtpServerClientFactory.Launch seam and VSTEST_CONNECTION_TIMEOUT -
so a save/restore in TestInitialize/TestCleanup could restore one class's value
while another class's test was still relying on its own. That would have flaked
in CI looking like a product bug.

Close a hole in the float range guard. (float)int.MaxValue rounds *up* to
2147483648f, so comparing a float directly against int.MaxValue let that value
through and the cast then saturated - precisely the plausible-looking wrong
answer the guard exists to reject. Widen to double before comparing, and extend
the regression test to cover it.

Capture ProcessId before the exit handshake instead of reading it afterwards,
when the process may already be gone.

Test fixes: TryExitDoesNotUseAnAlreadyCancelledRunToken was vacuous (it built a
cancelled token it never passed anywhere) and LaunchDefaultsToTheRealClientLauncher
asserted only non-null, which any delegate satisfies. Both now assert something
that fails if the behaviour regresses. Adds the missing mixed-selection case,
where only some tests carry a uid.

Also fixes a stale test-count comment and softens an overclaim in
MtpPureProject, which no test currently references.

Unit tests 142/142 across net11.0 and net481; MtpUnderVstestTests 16/16 on both
console axes.

Copilot-Session: cd0793f4-d530-44b6-b881-fed3be6aa52f

+++ src/Microsoft.TestPlatform.CrossPlatEngine/Client/MTP/MtpTestNodeConverter.cs
@@ -211,7 +211,11 @@ internal static class MtpTestNodeConverter
                 result = (int)d;
                 return true;
 
-            case float f when f is >= int.MinValue and <= int.MaxValue:
+            case float f
+                // (float)int.MaxValue rounds up to 2147483648f, so comparing a float against
+                // int.MaxValue directly lets that value through and the cast then saturates. Widen to
+                // double first so the bound is exact.
+                when (double)f is >= int.MinValue and <= int.MaxValue:
                 result = (int)f;
                 return true;
